rm cl add do
